Add PHP runtime inspection

+++ app/Console/Commands/NginxConfigCommand.php
@@ -43,6 +43,10 @@ class NginxConfigCommand extends Command
             $server = Server::find($this->argument('server'));
         }
 
+        if ($server->php && !is_file($server->php_fpm_socket)) {
+            throw new \RuntimeException('Selected PHP-FPM socket is missing: '.$server->php_fpm_socket);
+        }
+
         $config = view(sprintf('sample.nginx-%s', $server->template ?? 'default'), compact('server'));
         file_put_contents(storage_path(sprintf('blocks/%s.conf', $server->name)), $config);
